+++ packages/dash-table/src/dash-table/dash/DataTable.js
@@ -749,7 +749,8 @@ export const propTypes = {
             label: PropTypes.string.isRequired,
             value: PropTypes.oneOfType([
                 PropTypes.number,
-                PropTypes.string
+                PropTypes.string,
+                PropTypes.bool
             ]).isRequired
         })).isRequired
     })),
@@ -771,7 +772,8 @@ export const propTypes = {
             label: PropTypes.string.isRequired,
             value: PropTypes.oneOfType([
                 PropTypes.number,
-                PropTypes.string
+                PropTypes.string,
+                PropTypes.bool
             ]).isRequired
         })).isRequired
     })),
@@ -789,7 +791,8 @@ export const propTypes = {
                     label: PropTypes.string.isRequired,
                     value: PropTypes.oneOfType([
                         PropTypes.number,
-                        PropTypes.string
+                        PropTypes.string,
+                        PropTypes.bool
                     ]).isRequired
                 })).isRequired
             })